requirements.txt
pyTelegramBotAPI==3.2.0